Data processing script, and using processed data in webapp
