+++ package.json
@@ -6,6 +6,7 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
+    "processSynapseJSON": "npx ncc run data/processSynapseJSON.ts",
     "prettierPreCommit": "STAGED_AND_CHANGED_FILES=$(git diff HEAD --name-only --cached --diff-filter=d) && ([ -z \"$STAGED_AND_CHANGED_FILES\" ] && echo \"Nothing to prettify\" || (yarn run prettier --write $(echo $STAGED_AND_CHANGED_FILES) && git add -f $(echo $STAGED_AND_CHANGED_FILES)))",
     "prettierAll": "yarn run prettier --write $(git ls-files | grep '\\(.js\\|.ts\\|.scss\\|.css\\)')"
   },
@@ -68,6 +69,7 @@
     "@types/node": "^13.9.3",
     "@types/node-fetch": "^2.5.5",
     "@types/react": "^16.9.23",
+    "@vercel/ncc": "^0.28.6",
     "fork-ts-checker-webpack-plugin": "^6.1.0",
     "pre-commit": "^1.2.2",
     "typescript": "^3.8.3"
